pyproject.toml
[tool.poetry]
name = "PyRunner"
version = "0.1.0"
description = "A tool to run containerized CLI appication in batch via YAML config."
authors = ["Liam Wu <foo@bar.com>"]
readme = "README.md"
packages = [{include = "pyrunner"}]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
click = "^8.1.3"
rich = "^13.3.1"
pendulum = "^2.1.2"
docker = "^6.0.1"
pydantic = "^1.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pyrunner = 'pyrunner.app:run'
